requirements.txt
numpy>=1.24
scipy>=1.11
openai>=1.35
google-genai>=0.3.0
